% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_excel_from_template.R
\name{create_excel_from_template}
\alias{create_excel_from_template}
\title{Format data according to linked template}
\usage{
create_excel_from_template(
  x,
  file,
  sheet_name = "data",
  template_name = NULL,
  all_columns = FALSE,
  allow_new_cols = FALSE
)
}
\arguments{
\item{x}{Table to format (data frame or tibble)}

\item{file}{Location where to save the formatted xlsx}

\item{all_columns}{Whether to include all columns from the template in the
output file (empty columns will be filled in 'NA')}

\item{allow_new_cols}{Should columns with names NOT present in the template
be allowed? Setting this to true breaks the data standardization}
}
\value{
Nothing, this function is used for its side effects - saving a file
}
\description{
Format data according to linked template
}
