% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_structure.R
\name{create_structure}
\alias{create_structure}
\title{Create folder structure}
\usage{
create_structure()
}
\description{
This function creates folders inside your project folder to help organize your files
}
\examples{
create_structure()
}
