% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_table.R
\name{paste_table}
\alias{paste_table}
\alias{copy_table}
\title{Copy a table from R to clipboard or paste a table from clipboard into R.}
\source{
"Using the Windows Clipboard, or Passing Data Quickly From Excel to R and Back Again"
\url{https://www.r-bloggers.com/using-the-windows-clipboard-or-passing-data-quickly-from-excel-to-r-and-back-again/}
\url https://stackoverflow.com/questions/10959521/how-to-write-to-clipboard-on-ubuntu-linux-in-r
}
\usage{
paste_table()

copy_table(obj, size = 4096)
}
\arguments{
\item{obj}{Object that should be copied to clipboard}
}
\description{
Allows the user to easily and quickly copy or paste tables from or to R
in a format that is compatible with Microsoft Excel.
}
\details{
It really shines as a quick hacky tool for exploration, should never be used
in a script meant to be reproducible, for obvious reasons.
}
\examples{
# Import table in clipboard as an object
# Do not pass any argument
df <- paste_table()

# Copy a table to clipboard
copy_table(df)

}
