% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_with_error.R
\name{paste_with_error}
\alias{paste_with_error}
\title{Combine two vectors into a single vector with a separator}
\usage{
paste_with_error(value, error, sep = " ± ", digits = 2)
}
\arguments{
\item{value}{Numerical vector of value of a variable}

\item{error}{Numerical vector of error assocciated with variable}

\item{sep}{Separator for value and error values. Defaults to ±}

\item{digits}{Number of digits to numbers to. Defaults to 2}
}
\description{
Used to create tables for publication, where values are presented with their errors
The function will also deal with NAs, omitting the sep and error if error is NA
or returning a blank string the whole output if value is NA
}
\examples{
# This table makes to sense, but it shows how the function works
library(dplyr)
mutate(cars, example = paste_with_error(speed, dist))

}
