% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_structure.R
\name{research_project}
\alias{research_project}
\title{This function is called by Rstudio when a Research project rproj template is created
Defines folder structure and creates a README}
\usage{
research_project(path, ...)
}
\description{
This function is called by Rstudio when a Research project rproj template is created
Defines folder structure and creates a README
}
