% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{theme_custom}
\alias{theme_custom}
\title{A Custom ggplot2 Theme}
\usage{
theme_custom(serif = FALSE, grid = FALSE, axis = "xy")
}
\arguments{
\item{serif}{Changes all fonts to serif. Defaults to FALSE}

\item{grid}{Should plots contain grid lines at major breaks? Can be FALSE (default), "x", "y" or "xy"}
}
\description{
Sets up the basic look of my ggplot2 themes and abstracts some of the options
}
