% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_excel_from_template.R
\name{create_excel_from_template}
\alias{create_excel_from_template}
\title{Format data according to linked template}
\usage{
create_excel_from_template(
  x,
  file,
  sheet_name = "data",
  template_name = NULL,
  all_columns = FALSE,
  allow_new_cols = FALSE
)
}
\arguments{
\item{x}{Table to format (data frame or tibble)}

\item{file}{Location where to save the formatted xlsx}

\item{sheet_name}{Name of the sheet to save the data to}

\item{template_name}{Name of the template to use as a header}

\item{all_columns}{Should all columns from the template be included in the
output file, even if not present in data? (empty columns will be filled in 'NA')}

\item{allow_new_cols}{Should columns NOT present in the template
be allowed?}
}
\value{
Nothing, this function is used for its side effects - saving a file
}
\description{
Format data according to linked template
}
