% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_xlsx.R
\name{df_to_xlsx}
\alias{df_to_xlsx}
\title{Add Dataframe to Excel File}
\usage{
df_to_xlsx(excel_file, sheet_name, df, merge_cols = NULL)
}
\arguments{
\item{excel_file}{Location of excel file where data.frame will be added}

\item{sheet_name}{Name of the sheet inside the file -will replace sheets with same name-}

\item{df}{Data.frame to be added to file}

\item{merge_cols}{Character vector with names of columns. Categories on assumed to be nested (breaks from columns on the left will be carried on to columns on the right)}
}
\description{
This function adds a data.frame to an excel file, formatting the header. Also allows to select columns where rows with equal values are merged.
}
\examples{
df_to_xlsx(excel_file = "./excel.xlsx", sheet_name = "information", df = info_df, merge_cols = c("Continent", "Country"))
}
