% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_xlsx.R
\name{df_to_xlsx}
\alias{df_to_xlsx}
\title{Add Dataframe to Excel File}
\usage{
df_to_xlsx(
  excel_file,
  sheet_names,
  dfs,
  auto_size_cols = TRUE,
  merge_cols = NULL,
  nested_merge = TRUE
)
}
\arguments{
\item{excel_file}{Location of excel file where data.frame will be added}

\item{auto_size_cols}{Should column width be changed based on their contents? TRUE or FALSE, defaults to TRUE}

\item{merge_cols}{Names of columns that should be merged vertically when they have the same value. Default is no merging, expects a character vector. Currently only works if you pass a single df}

\item{nested_merge}{Should merging be done with columns on the right being nested inside the ones on the left? TRUE or FALSE}

\item{sheet_name}{Names of the sheet inside the file -will replace sheets with same name-}

\item{df}{List of data.frame to be added to file}
}
\value{
Excel file with data.frame added
}
\description{
This function adds a data.frame to an excel file, formatting the header. Also allows to select columns where rows with equal values are merged.
}
\examples{
df_to_xlsx(excel_file = "./excel.xlsx", sheet_name = "information", df = info_df, merge_cols = c("Continent", "Country"))
}
